backend/requirements.txt
flask==3.1.1
mysql-connector-python==9.3.0
pymongo==4.13.0
gunicorn